requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9; platform_system != "Windows" or python_version < "3.13"
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-telegram-bot>=21.0
cryptography>=41.0.0
icalendar>=5.0.0
pytest==7.4.3
pytest-flask==1.3.0

